Revert "chore(litmus-portal): Frontend functionality for resend invitation, cancel invitation and remove member added (#2207)"
This reverts commit 2dc67e4ccb101f1cd3c4c926f415a18c65bb4b27.
