+++ litmus-portal/frontend/src/graphql/mutations.ts
@@ -34,12 +34,6 @@ export const SEND_INVITE = gql`
   }
 `;
 
-export const REMOVE_INVITATION = gql`
-  mutation RemoveInvitation($data: MemberInput!) {
-    removeInvitation(member: $data)
-  }
-`;
-
 export const ACCEPT_INVITE = gql`
   mutation accept($member: MemberInput!) {
     acceptInvitation(member: $member)
