+++ litmus-portal/frontend/src/views/Settings/TeammingTab/index.tsx
@@ -1,9 +1,12 @@
 import { useQuery } from '@apollo/client/react/hooks';
 import {
+  Avatar,
   createStyles,
   FormControl,
+  IconButton,
   InputAdornment,
   InputLabel,
+  Menu,
   MenuItem,
   Paper,
   Select,
@@ -21,6 +24,7 @@ import {
   withStyles,
 } from '@material-ui/core';
 import SearchIcon from '@material-ui/icons/Search';
+import moment from 'moment';
 import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { GET_USER } from '../../../graphql';
@@ -30,10 +34,10 @@ import {
   Member,
 } from '../../../models/graphql/user';
 import { RootState } from '../../../redux/reducers';
+import userAvatar from '../../../utils/user';
 import Invitation from './Invitation';
 import InviteNew from './InviteNew';
 import useStyles from './styles';
-import TableData from './tableData';
 
 // StyledTableCell used to create custom table cell
 const StyledTableCell = withStyles((theme: Theme) =>
@@ -84,7 +88,7 @@ const TeammingTab: React.FC = () => {
     role: 'all',
   });
 
-  // const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
 
   // logic for displaying the team members of the user
   let memberList: Member[] = [];
@@ -125,6 +129,17 @@ const TeammingTab: React.FC = () => {
         if (filters.role === 'Viewer') return dataRow.role === 'Viewer';
         return dataRow.role === 'Owner';
       });
+
+  // for closing the menu option
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
+
+  // Function to display date in format Do MMM,YYYY Hr:MM AM/PM
+  const formatDate = (date: string) => {
+    const day = moment(date).format('Do MMM, YYYY LT');
+    return day;
+  };
 
   return (
     <div>
@@ -234,7 +249,132 @@ const TeammingTab: React.FC = () => {
                           key={row.name}
                           className={classes.TR}
                         >
-                          <TableData index={index} row={row} />
+                          <TableCell
+                            className={classes.firstTC}
+                            component="th"
+                            scope="row"
+                          >
+                            <div className={classes.firstCol}>
+                              <Avatar
+                                data-cy="avatar"
+                                alt="User"
+                                className={classes.avatarBackground}
+                              >
+                                {row?.name
+                                  ? userAvatar(row.name)
+                                  : userAvatar(row.name)}
+                              </Avatar>
+                              {row.name}
+                            </div>
+                          </TableCell>
+                          <TableCell className={classes.otherTC}>
+                            {row.role}
+                            {row.role === 'editor' || row.role === 'viewer' ? (
+                              <>
+                                <IconButton
+                                  disabled
+                                  aria-label="more"
+                                  aria-controls="long-menu"
+                                  aria-haspopup="true"
+                                  onClick={(event) => {
+                                    setAnchorEl(event.currentTarget);
+                                  }}
+                                  className={classes.optionBtn}
+                                >
+                                  <img
+                                    src="./icons/right-arrow.svg"
+                                    alt="more"
+                                  />
+                                </IconButton>
+                                <Menu
+                                  keepMounted
+                                  open={Boolean(anchorEl)}
+                                  id="long-menu"
+                                  anchorEl={anchorEl}
+                                  onClose={handleClose}
+                                >
+                                  <MenuItem
+                                    value={index}
+                                    onClick={() => {
+                                      setAnchorEl(null);
+                                    }}
+                                    className={classes.menuOpt}
+                                  >
+                                    <div className={classes.menuDiv}>
+                                      <div>
+                                        <Typography
+                                          className={classes.menuHeader}
+                                        >
+                                          <strong>Editor</strong>
+                                        </Typography>
+                                      </div>
+                                      <div>
+                                        <Typography
+                                          className={classes.menuDesc}
+                                        >
+                                          Can make changes in the project
+                                        </Typography>
+                                      </div>
+                                    </div>
+                                  </MenuItem>
+                                  <MenuItem
+                                    value={index}
+                                    onClick={() => {
+                                      setAnchorEl(null);
+                                    }}
+                                    className={classes.menuOpt}
+                                  >
+                                    <div className={classes.menuDiv}>
+                                      <div>
+                                        <Typography
+                                          className={classes.menuHeader}
+                                        >
+                                          <strong>Viewer</strong>
+                                        </Typography>
+                                      </div>
+                                      <div>
+                                        <Typography
+                                          className={classes.menuDesc}
+                                        >
+                                          Can make changes in the project
+                                        </Typography>
+                                      </div>
+                                    </div>
+                                  </MenuItem>
+                                </Menu>
+                              </>
+                            ) : (
+                              <></>
+                            )}
+                          </TableCell>
+                          <TableCell className={classes.otherTC}>
+                            {row.email}
+                          </TableCell>
+                          <TableCell className={classes.otherTC}>
+                            <div className={classes.dateDiv}>
+                              <img
+                                className={classes.calIcon}
+                                src="./icons/calendarIcon.svg"
+                                alt="calendar"
+                              />
+                              {formatDate(row.joined_at)}
+                            </div>
+                          </TableCell>
+
+                          <TableCell
+                            className={classes.otherTC}
+                            key={row.user_name}
+                          >
+                            {/*  <DelUser
+                              handleTable={() => {}}
+                              tableDelete={false}
+                              teammingDel
+                              disabled
+                            /> */}
+                            <IconButton disabled onClick={() => {}}>
+                              <img alt="delete" src="./icons/bin-grey.svg" />
+                            </IconButton>
+                          </TableCell>
                         </TableRow>
                       ))
                   ) : (
